refactor(m-api): DeviceStatusManager 추상화 제거
다른 기술을 이용할 일 없으며 여러 가지 버전이 필요할만한 기능이 아니라고 판단했습니다.
